Move test_factory_creation out of TestHelpers.hpp
It needs ParseOptions.hpp, which is slow to compile.
