+++ tests/Unit/PointwiseFunctions/MathFunctions/Test_Gaussian.cpp
@@ -6,6 +6,7 @@
 #include "Parallel/RegisterDerivedClassesWithCharm.hpp"
 #include "PointwiseFunctions/MathFunctions/Gaussian.hpp"
 #include "tests/Unit/PointwiseFunctions/MathFunctions/TestMathHelpers.hpp"
+#include "tests/Unit/TestFactoryCreation.hpp"
 #include "tests/Unit/TestHelpers.hpp"
 
 SPECTRE_TEST_CASE("Unit.PointwiseFunctions.MathFunctions.Gaussian",
