+++ tests/Unit/Time/TimeSteppers/Test_RungeKutta3.cpp
@@ -4,6 +4,7 @@
 #include "Parallel/PupStlCpp11.hpp"
 #include "Parallel/RegisterDerivedClassesWithCharm.hpp"
 #include "Time/TimeSteppers/RungeKutta3.hpp"
+#include "tests/Unit/TestFactoryCreation.hpp"
 #include "tests/Unit/TestHelpers.hpp"
 #include "tests/Unit/Time/TimeSteppers/TimeStepperTestUtils.hpp"
